requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Machine Learning dependencies
torch==2.1.0
transformers==4.35.0
optimum[onnxruntime]==1.14.0
onnxruntime==1.16.0

# Additional dependencies that might be needed
numpy==1.24.3
google-genai
motor
